pyproject.toml
[build-system]
requires = ["setuptools>=64.0", "wheel", "tomli"]
build-backend = "build_backend"
backend-path = ["."]

[project]
name = "sixbirds-cosmo"
version = "0.0.0"
requires-python = ">=3.10"
dependencies = [
  "numpy",
  "scipy",
  "pandas",
  "matplotlib",
  "pyyaml",
  "astropy",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-cov",
]
lss = [
  "sacc",
  "pyccl",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
